+++ src/main/java/br/ufrpe/minhacampanha/dao/Donativo_produtoDAO.java
@@ -54,7 +54,7 @@ public class Donativo_produtoDAO {
 			while (resultSet.next()){
 				
 				Donativo_produto produto = new Donativo_produto();
-				produto.setCodigo(resultSet.getLong("id"));
+				produto.setCodigo(resultSet.getInt("id"));
 				//produto.setDataDoacao(resultSet.getDate("data_doacao"));
 				produto.setId_campanha(resultSet.getInt("id_campanha"));
 				produto.setId_inst_doadora(resultSet.getInt("id_inst_doadora"));
